refactor: api ts 적용
