--- src/api/index.js
@@ -1,21 +1,0 @@
-import axios from "axios";
-
-const baseURL = process.env.REACT_APP_API_URL;
-const token = process.env.REACT_APP_TOKEN;
-
-const baseInstance = axios.create({
-  baseURL,
-  headers: {
-    Authorization: `Bearer ${token}`,
-  },
-});
-
-baseInstance.interceptors.response.use(({ data }) => data);
-
-const apiRequest = {
-  get: (url, request) => baseInstance.get(url, request),
-  delete: (url, request) => baseInstance.delete(url, request),
-  post: (url, data, config) => baseInstance.post(url, data, config),
-};
-
-export default apiRequest;
