+++ src/api/todo.ts
@@ -1,6 +1,6 @@
-import apiRequest from "./index";
+import apiRequest from './index';
 
-const RESOURCE = "/todos";
+const RESOURCE = '/todos';
 
 export const getTodoList = async () => {
   try {
@@ -8,26 +8,26 @@ export const getTodoList = async () => {
 
     return response;
   } catch (error) {
-    throw new Error("API getTodoList error");
+    throw new Error('API getTodoList error');
   }
 };
 
-export const createTodo = async (data) => {
+export const createTodo = async (data: { title: string }) => {
   try {
     const response = await apiRequest.post(`${RESOURCE}`, data);
 
     return response;
   } catch (error) {
-    throw new Error("API createTodo error");
+    throw new Error('API createTodo error');
   }
 };
 
-export const deleteTodo = async (id) => {
+export const deleteTodo = async (id: number) => {
   try {
     const response = await apiRequest.delete(`${RESOURCE}/${id}`);
 
     return response;
   } catch (error) {
-    throw new Error("API deleteTodo error");
+    throw new Error('API deleteTodo error');
   }
 };
